impl WalletRepository

+++ src/infrastructure/schema.rs
@@ -7,14 +7,11 @@ table! {
 table! {
     wallet (id) {
         id -> Varchar,
-        deposit -> Int8,
+        deposit -> BigInt,
         currency -> Nullable<Varchar>,
     }
 }
 
 joinable!(wallet -> currency (currency));
 
-allow_tables_to_appear_in_same_query!(
-    currency,
-    wallet,
-);
+allow_tables_to_appear_in_same_query!(currency, wallet,);
